+++ eft-dma-radar/Tarkov/EFTPlayer/Player.cs
@@ -307,6 +307,8 @@ namespace eft_dma_radar.Tarkov.EFTPlayer
         /// </summary>
         public virtual ulong RotationAddress { get; }
         public virtual float ZoomLevel { get; set; }
+        public virtual ulong PWA { get; set; }
+        public bool IsAiming { get; set; }
 
         #endregion
 
@@ -456,6 +458,8 @@ namespace eft_dma_radar.Tarkov.EFTPlayer
             if (isActive)
             {
                 this.SetAlive();
+                this.IsAiming = Hands.IsAiming;
+                this.ZoomLevel = Hands.ZoomLevel;
             }
             else if (this.IsAlive) // Not in list, but alive
             {
